+++ rustyhack_client/src/screens/side_status_bar.rs
@@ -1,5 +1,7 @@
+use crate::client_consts::DEFAULT_BG_COLOUR;
 use console_engine::screen::Screen;
 use console_engine::ConsoleEngine;
+use crossterm::style::Color;
 use rustyhack_lib::ecs::item::get_item_name;
 use rustyhack_lib::ecs::player::Player;
 
@@ -17,6 +19,7 @@ pub(crate) fn draw(player: &Player, console: &ConsoleEngine, viewport_width: u32
     let str_string = "Str: ".to_owned() + &player.stats.str.to_string();
     let dex_string = "Dex: ".to_owned() + &player.stats.dex.to_string();
     let con_string = "Con: ".to_owned() + &player.stats.con.to_string();
+    let stat_points_string = "Stat (u)p available!";
 
     let gold_string = "Gold: ".to_owned() + &player.inventory.gold.to_string();
 
@@ -51,21 +54,26 @@ pub(crate) fn draw(player: &Player, console: &ConsoleEngine, viewport_width: u32
 
     let inventory_title_string = "Inventory:";
 
-    screen.print(1, 0, &player.player_details.player_name);
-    screen.print(1, 1, &lvl_string);
-    screen.print(1, 2, &exp_string);
-    screen.print(1, 3, &exp_next_string);
-    screen.print(1, 5, &hp_string);
-    screen.print(1, 6, &str_string);
-    screen.print(1, 7, &dex_string);
-    screen.print(1, 8, &con_string);
-    screen.print(1, 10, &gold_string);
-    screen.print(1, 12, equipped_title_string);
-    screen.print(1, 13, &weapon_string);
-    screen.print(1, 14, &armour_string);
-    screen.print(1, 16, inventory_title_string);
+    let mut y = 0;
+    screen.print(1, y, &player.player_details.player_name);
+    screen.print(1, y + 1, &lvl_string);
+    screen.print(1, y + 2, &exp_string);
+    screen.print(1, y + 3, &exp_next_string);
+    screen.print(1, y + 5, &hp_string);
+    screen.print(1, y + 6, &str_string);
+    screen.print(1, y + 7, &dex_string);
+    screen.print(1, y + 8, &con_string);
+    if player.stats.stat_points > 0 {
+        screen.print_fbg(1, y + 9, stat_points_string, Color::Cyan, DEFAULT_BG_COLOUR);
+        y += 1;
+    }
+    screen.print(1, y + 10, &gold_string);
+    screen.print(1, y + 12, equipped_title_string);
+    screen.print(1, y + 13, &weapon_string);
+    screen.print(1, y + 14, &armour_string);
+    screen.print(1, y + 16, inventory_title_string);
 
-    let mut line_count = 17;
+    let mut line_count = y + 17;
     for item in &player.inventory.carried {
         let item_text = get_item_name(item);
         screen.print(1, line_count, &item_text);