+++ rustyhack_lib/src/ecs/monster.rs
@@ -54,6 +54,7 @@ impl Default for Monster {
                 str: 1.0,
                 dex: 1.0,
                 con: 1.0,
+                stat_points: 0,
                 level: 1,
                 exp: 0,
                 exp_next: 1000,